refactor: 도메인은 Model suffix 추가
